fix to activity test

+++ 03-Development/dto/src/test/java/co/shinetech/dto/ActivityTest.java
@@ -24,7 +24,7 @@ public class ActivityTest {
 		Profile prof = new Profile(400,"Teacher");
 		User teacher = new User(300,prof);
 		LocalDateTime ldt = LocalDateTime.now().plusHours(1);
-		Activity a = new Activity((long)39,"Joel",LocalDateTime.now(),ldt,"class1",teacher);
+		Activity a = new Activity(39,"Joel",LocalDateTime.now(),ldt,"class1",teacher);
 		assertNotNull(a);
 	}
 
@@ -33,8 +33,8 @@ public class ActivityTest {
 		Profile prof = new Profile(400,"Teacher");
 		User teacher = new User(300,prof);
 		LocalDateTime ldt = LocalDateTime.now().plusHours(1);
-		Activity a1 = new Activity((long)39,"Joel",LocalDateTime.now(),ldt,"class1",teacher);
-		Activity a2 = new Activity((long)39,"Joel",LocalDateTime.now(),ldt,"class1",teacher);
+		Activity a1 = new Activity(39,"Joel",LocalDateTime.now(),ldt,"class1",teacher);
+		Activity a2 = new Activity(39,"Joel",LocalDateTime.now(),ldt,"class1",teacher);
 		System.out.println(a1.toString());
 		System.out.println(a2.toString());
 		assertEquals(a1.hashCode(),a2.hashCode());
